Add SQL queries to delete matches and to add participants to matches

+++ dbfunctions/modify_match.php
@@ -1,8 +1,8 @@
 <?php
 
 // match_data: participants, result, teams
+// TODO(lucas): Put participants in match_data, and call match_add_participants from here!
 // NOTE(lucas): Team 1 is the team that created the match!
-// Returns true for success, false for failure
 function match_create($link, $match_data) {
 	if (!$match_data) {
 		return false;
@@ -19,31 +19,52 @@ function match_create($link, $match_data) {
 	return false;
 }
 
-// Exclude participants in a team that haven't accepted their invite yet, just in case.
 function match_add_participants($link, $id, $participants, $team) {
 	$id = mysqli_real_escape_string($link, $id);
 	$team = mysqli_real_escape_string($link, $team);
-	$query = '';
+	$query = 'DELETE FROM MatchParticipation WHERE Matches = ' . $id . ' AND Team = "' . $team . '";';
+	if (!mysqli_query($link, $query)) {
+		return false;
+	}
+	$status = true;
+	for ($i = 0; $i < count($participants); ++$i) {
+		$user_id = mysqli_real_escape_string($participants[$i]);
+		$query = 'INSERT INTO MatchParticipation (User, Matches, Team) VALUES(' . $user_id . ', ' . $id . ', "' . $team . '")';
+		if (!mysqli_query($link, $query)) {
+			$status = false;
+		}
+	}
+	return $status;
 }
 
-function match_modify($link, $id, $match_data, $team) {
+// match_data: participants, result
+// TODO(lucas): Put participants in match_data, and call match_add_participants from here!
+// NOTE(lucas): Modify results, team participants and then toggle the value of Team2ShouldVerify to pass on the control to the other team
+function match_modify($link, $id, $match_data) {
 	$id = mysqli_real_escape_string($link, $id);
-	$team = mysqli_real_escape_string($link, $team);
+	$result = mysqli_real_escape_string($link, $match_data["result"]);
+	$query = 'UPDATE Matches SET Result = "' . $result . '" Team2ShouldVerify = !Team2ShouldVerify WHERE Id = ' . $id . ';';
+	if (!mysqli_query($link, $query)) {
+		return false;
+	}
+	// TODO(lucas): Participants
+	return false;
+}
+
+function match_verify($link, $id) {
+	$id = mysqli_real_escape_string($link, $id);
 
 	return false;
 }
 
-function match_verify($link, $id, $team) {
+function match_delete($link, $id) {
 	$id = mysqli_real_escape_string($link, $id);
-	$team = mysqli_real_escape_string($link, $team);
-
+	$query = 'DELETE FROM MatchParticipation WHERE Matches = ' . $id . ';';
+	if (mysqli_query($link, $query)) {
+		$query = 'DELETE FROM Matches WHERE Id = ' . $id . ';';
+		if (mysqli_query($link, $query)) {
+			return true;
+		}
+	}
 	return false;
-}
-
-// TODO(lucas): Unlink match participants!
-function match_delete($link, $id, $team) {
-	$id = mysqli_real_escape_string($link, $id);
-	$team = mysqli_real_escape_string($link, $team);
-	$query = 'DELETE FROM Matches WHERE Id = ' . $id . ';';
-	return true;
 }
